+++ aksel.nav.no/website/pages/eksempler/popover/no-offset.tsx
@@ -21,7 +21,6 @@ const Example = () => {
         onClose={() => setOpenState(false)}
         anchorEl={anchorEl}
         offset={0}
-        arrow={false}
         placement="bottom"
       >
         <Popover.Content>offset = 0</Popover.Content>