+++ @navikt/aksel-icons/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/aksel-icons
 
+## 7.34.0
+
+### Minor Changes
+
+- Icons: :tada: New icons `PhoneSlash` and `BellSlash`. ([#4310](https://github.com/navikt/aksel/pull/4310))
+
+## 7.33.5
+
+## 7.33.4
+
+## 7.33.3
+
 ## 7.33.2
 
 ## 7.33.1