+++ @navikt/aksel/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @navikt/aksel
 
+## 7.34.0
+
+### Patch Changes
+
+- Updated dependencies [[`39cf344`](https://github.com/navikt/aksel/commit/39cf344eeb0c01560054e47218bcf9ce7c1bbf35), [`39cf344`](https://github.com/navikt/aksel/commit/39cf344eeb0c01560054e47218bcf9ce7c1bbf35), [`39cf344`](https://github.com/navikt/aksel/commit/39cf344eeb0c01560054e47218bcf9ce7c1bbf35), [`efebc5f`](https://github.com/navikt/aksel/commit/efebc5fbbb3e955df5050086aef68812052a837d), [`39cf344`](https://github.com/navikt/aksel/commit/39cf344eeb0c01560054e47218bcf9ce7c1bbf35)]:
+  - @navikt/ds-css@7.34.0
+  - @navikt/ds-tokens@7.34.0
+
+## 7.33.5
+
+### Patch Changes
+
+- Updated dependencies [[`5cfff15`](https://github.com/navikt/aksel/commit/5cfff15c1c367a61d23ba3498a350669789dad81)]:
+  - @navikt/ds-css@7.33.5
+  - @navikt/ds-tokens@7.33.5
+
+## 7.33.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @navikt/ds-css@7.33.4
+  - @navikt/ds-tokens@7.33.4
+
+## 7.33.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @navikt/ds-css@7.33.3
+  - @navikt/ds-tokens@7.33.3
+
 ## 7.33.2
 
 ### Patch Changes